Letters Collection
The Postmaster wants to write a program to answer the queries regarding letter collection in a city.
A city is represented as a matrix mat of size n*m.
Each cell represents a house and contains letters which are denoted by a number in the cell. 
The program should answer q queries which are of following types:
1 i j : To sum all the letters which are at a 1-hop distance from the cell (i,j) in any direction
2 i j : To sum all the letters which are at a 2-hop distance from the cell (i,j) in any direction 
The queries are given in a 2D matrix queries[][].
In one hop distance postmaster can go to any of [(i-1,j-1), (i-1,j), (i-1,j+1), (i,j-1), (i,j+1), (i+1,j-1), (i+1,j), (i+1,j+1)] from (i,j).